Extra nginx exposure
think this already landed and this is probably moot.

Expose the analysis files as "raw-analysis" and the "repo-files" and
"objdir-files" under "file-lists".  This is intended to make debugging
investigations easier as well as to enable extraction of per-file and
per-directory information without having to scrape the HTML output of
output-file.