Improve graphviz rendering to support clusters as edge targets.
Very much informed by a variety of good advice at
https://stackoverflow.com/questions/2012036/graphviz-how-to-connect-subgraphs

Although it probably indicates a logic error in whatever is
drawing the diagram, it's currently possible to create diagrams where
an edge targets a cluster.  For example 'method' can get confused when
a type is clicked on and it can go and diagram the all the related
constructors thanks to super-classes.  Or maybe it's confused by the
type hierarchy.  Either way, the sub-classes can want to reference
the superclass, and if the super-class has also a method that gets
displayed, this will cause the class to become a cluster.

These changes make the diagram compound so that we can use lhead/ltail
to have the arrows stop at the cluster's edge.  We also introduce an
invisible internal node which is what the edges are actually pointing
at.  (We still need to have edges that go to/from something inside the
cluster.)  Since we already were using the cluster name as the edgeInId /
edgeOutId, it was reasonable straightforward to just have those name the
sentinel node.