Use the livegrep revision of the submodule from upstream.
When introducing the livegrep submodule, provisioning did not change,
but because of build problems, we need to update and I'm addressing
that.

Disclaimer: I don't think I actually needed to upgrade livegrep;
the problem is that I forgot that during the bazel build the build
seems to hang because it's waiting to try and get something involving
boost from sourceforge.  Unfortunately, this may mean waiting for a
few minutes while nothing seems to happen.  It seems good to be
generally up-to-date with livegrep, though, so I'm keeping this change.

Note that this change does potentially leave some loose ends for AWS
provisioning which currently will not properly honor the checked out
submodule revision, but that's not a current concern.